+++ src/OwnerCollection.java
@@ -14,7 +14,7 @@ public class OwnerCollection {
     public HashSet<Owner> getOwners(){
         return this.owners;
     }
-    public Owner getOwnerByCreditCardNumber(int creditCardNumber){
+    public Owner getOwnerByCreditCardNumber(long creditCardNumber){
         for (Owner owner : this.owners){
             if (owner.getCreditCard().getNumber() == creditCardNumber){
                 return owner;
@@ -23,7 +23,7 @@ public class OwnerCollection {
         return null;
     }
     public void UpdateOwnership(Owner newOwner){
-        int CreditCardNumber = newOwner.getCreditCard().getNumber();
+        long CreditCardNumber = newOwner.getCreditCard().getNumber();
         for (Owner owner : this.owners){
             if (owner.getCreditCard().getNumber() == CreditCardNumber){
                 this.owners.remove(owner);
